docs/source/requirements.txt
sphinx-hoverxref
sphinx-copybutton
sphinx-design
pydata-sphinx-theme